+++ packages/core/src/components/EmojiPicker.tsx
@@ -276,7 +276,7 @@ const EmojiPicker = ({
         value={inputValue}
         height="s"
         onChange={(e) => setInputValue(e.target.value)}
-        hasPrefix={<Icon size="s" onBackground="neutral-weak" name="search" />}
+        prefix={<Icon size="s" onBackground="neutral-weak" name="search" />}
         aria-label="Search emojis"
       />
 
@@ -313,8 +313,8 @@ const EmojiPicker = ({
       {!searchQuery && (
         <SegmentedControl
           buttons={categoryButtons}
-          onToggle={handleCategoryChange}
-          defaultSelected={activeCategory}
+          onChange={handleCategoryChange}
+          defaultValue={activeCategory}
           fillWidth
         />
       )}